Add files via upload
AppleStore.csv contains the values for the different categories, such as price, average ratings, etc. the description files are separate as they are large in size and contain the description of the various apps. In the .sql file these descriptions are merged with the data in the AppleStore.csv file.